Implement basic online multiplayer
For now, I had to get rid of most of the game elements to make it easier to implement this (removing player HUD, most colors, balls/ball physics). I had to restructure/refactor most of the code to get a good setup going. Now that it works, I'll start re-implementing those features.